Completed the file structure.
I'm now positive that this is how I'm going to have the functions set up. In terms of function return type and parameters, I think it is set. I still haven't added the DMA aspect yet, but that can wait for now. I'll move on to the menu() function before that.