Menu loop created
I would love to say it is complete, but not quite. The accepting of the inputs isn't complete. There are cases where it sold accept 3 or 4. Right now, the code only works with 3. Other than that, each command will call the respective function to carry out it's operations. These functions aren't completed either.